move logging setup to new package
