+++ internal/logging/logging_test.go
@@ -1,4 +1,4 @@
-package main
+package logging
 
 import (
 	"context"
@@ -8,6 +8,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	"k8s.io/klog/v2"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
 
@@ -16,12 +17,13 @@ func TestSetupLogging(t *testing.T) {
 	r, w, _ := os.Pipe()
 	os.Stderr = w
 
-	setupLogging()
+	SetupLogging()
 	log := ctrl.LoggerFrom(context.Background())
 	log = log.WithValues("tempo", "simplest")
 	log.V(1).Info("a test debug message")
 	log.Info("a test info message")
 	log.Error(errors.New("test error"), "a test error occurred")
+	klog.Info("test from klog")
 
 	err := w.Close()
 	require.NoError(t, err)
@@ -31,4 +33,5 @@ func TestSetupLogging(t *testing.T) {
 	require.Regexp(t, `{"level":"info","ts":".+","msg":"a test info message","tempo":"simplest"}`, string(output))
 	require.NotRegexp(t, `{"level":"debug",.+}`, string(output))
 	require.Regexp(t, `{"level":"error","ts":".+","msg":"a test error occurred","tempo":"simplest","error":"test error","stacktrace":`, string(output))
+	require.Regexp(t, `{"level":"info","ts":".+","msg":"test from klog"}`, string(output))
 }
